♻️ refactor: 重构图片上传逻辑 将图片上传功能抽取为独立的工具函数，提高代码复用性。主要更改： 创建 uploadImage.js 统一处理图片上传逻辑 重构 link.js，移除重复代码 更新 MJ_Painting.js 使用新的 uploadImage 函数 优化错误处理和日志记录
