+++ apps/link.js
@@ -1,8 +1,8 @@
 import plugin from '../../../lib/plugins/plugin.js'
-import fetch from 'node-fetch'
-import fs from 'fs'
-import path from 'path'
 import Config from '../components/Config.js'
+import { parseSourceImg } from '../utils/getImg.js'
+import { uploadImage } from '../utils/uploadImage.js'
+import fetch from 'node-fetch'
 
 export class LinkPlugin extends plugin {
     constructor() {
@@ -27,46 +27,6 @@ export class LinkPlugin extends plugin {
                 }
             ]
         })
-
-        this.tempDir = './temp'
-        if (!fs.existsSync(this.tempDir)) {
-            fs.mkdirSync(this.tempDir)
-        }
-    }
-
-    /**
-     * 获取图片链接
-     * @param {object} e 事件对象
-     * @returns {Array} 图片链接数组
-     */
-    async getImg(e) {
-        if (!this.e.isMaster) return false;
-
-        if (e.at && !e.source) {
-            e.img = [`https://q1.qlogo.cn/g?b=qq&s=0&nk=${e.at}`];
-        }
-        if (e.source) {
-            let reply;
-            let seq = e.isGroup ? e.source.seq : e.source.time;
-            if (e.adapter === 'shamrock') {
-                seq = e.source.message_id;
-            }
-            if (e.isGroup) {
-                reply = (await e.group.getChatHistory(seq, 1)).pop()?.message;
-            } else {
-                reply = (await e.friend.getChatHistory(seq, 1)).pop()?.message;
-            }
-            if (reply) {
-                let i = [];
-                for (let val of reply) {
-                    if (val.type === 'image') {
-                        i.push(val.url);
-                    }
-                }
-                e.img = i;
-            }
-        }
-        return e.img;
     }
 
     /**
@@ -74,96 +34,27 @@ export class LinkPlugin extends plugin {
      * @param {object} e 事件对象
      */
     async zhil(e) {
-        if (!this.e.isMaster) return false;
-        const config = Config.getConfig()
-        const domain = config.link_domain
-
         console.log('收到命令:', e.msg);
 
-        const imgUrls = await this.getImg(e);
-
-        if (!imgUrls || imgUrls.length === 0) {
+        await parseSourceImg(e)
+        
+        if (!e.img || e.img.length === 0) {
             await this.reply('❌未找到有效的图片链接。');
             return true;
         }
 
-        const imgUrl = imgUrls[0];
+        const imgUrl = e.img[0];
         console.log('匹配到的图片链接:', imgUrl);
 
-        let tempFilePath = null;
-
         try {
-            // 添加超时控制
-            const controller = new AbortController();
-            const timeout = setTimeout(() => controller.abort(), 30000); // 30秒超时
-
-            const response = await fetch(imgUrl, { 
-                signal: controller.signal,
-                timeout: 30000 
-            });
-            clearTimeout(timeout);
-
-            if (!response.ok) {
-                throw new Error(`获取图片失败: ${response.status} ${response.statusText}`);
-            }
-
-            const arrayBuffer = await response.arrayBuffer();
-            const imgBuffer = Buffer.from(arrayBuffer);
-
-            tempFilePath = path.join(this.tempDir, `image_${Date.now()}.jpg`);
-            fs.writeFileSync(tempFilePath, imgBuffer);
-
-            const boundary = '----WebKitFormBoundary' + Math.random().toString(36).slice(2);
-            let formBody = '';
-
-            formBody += `--${boundary}\r\n`;
-            formBody += 'Content-Disposition: form-data; name="file"; filename="image.jpg"\r\n';
-            formBody += 'Content-Type: image/jpeg\r\n\r\n';
-            formBody += Buffer.from(fs.readFileSync(tempFilePath)).toString('binary');
-            formBody += `\r\n--${boundary}--\r\n`;
-
-            // 添加新的超时控制
-            const uploadController = new AbortController();
-            const uploadTimeout = setTimeout(() => uploadController.abort(), 30000);
-
-            const uploadResponse = await fetch(`${domain}/upload.php`, {
-                method: 'POST',
-                body: Buffer.from(formBody, 'binary'),
-                headers: {
-                    'Content-Type': `multipart/form-data; boundary=${boundary}`
-                },
-                signal: uploadController.signal,
-                timeout: 30000
-            });
-            clearTimeout(uploadTimeout);
-
-            if (!uploadResponse.ok) {
-                throw new Error(`上传失败: ${uploadResponse.status} ${uploadResponse.statusText}`);
-            }
-
-            const uploadResult = await uploadResponse.json();
-            if (uploadResult.code !== 200) {
-                await this.reply(`失败：${uploadResult.msg}`);
-            } else if (uploadResult.img) {
-                const fixedUrl = uploadResult.img.replace(/([^:])\/+/g, '$1/');
-                await this.reply(fixedUrl);
-            }
-
+            const uploadedUrl = await uploadImage(imgUrl);
+            await this.reply(uploadedUrl);
         } catch (err) {
-            console.error('失败:', err);
+            console.error('上传失败:', err);
             if (err.name === 'AbortError') {
                 await this.reply('请求超时，请稍后重试');
             } else {
-                await this.reply('失败：' + err.message);
-            }
-        } finally {
-            // 清理临时文件
-            if (tempFilePath && fs.existsSync(tempFilePath)) {
-                try {
-                    fs.unlinkSync(tempFilePath);
-                } catch (err) {
-                    console.error('删除临时文件失败:', err);
-                }
+                await this.reply('上传失败：' + err.message);
             }
         }
 
@@ -175,7 +66,6 @@ export class LinkPlugin extends plugin {
      * @param {object} e 事件对象
      */
     async deleteLink(e) {
-        if (!this.e.isMaster) return false;
         const config = Config.getConfig()
         const domain = config.link_domain
 
@@ -233,8 +123,6 @@ export class LinkPlugin extends plugin {
      * @param {object} e 事件对象
      */
     async setDomain(e) {
-        if (!this.e.isMaster) return false;
-
         let domain = e.msg.replace(/^#设置直链域名/, '').trim()
         
         if (!domain) {
